requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
opencv-python
numpy
Pillow
--index-url https://pypi.org/simple
--only-binary :all: 